refactor: rename .hpp to .h for header files
