+++ src/gui/WindowManager.h
